# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Task API",
        "description": "API para gestión de proyectos y tareas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/projects": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Obtener lista de proyectos",
                "operationId": "07f276c6e629646d087ee4b728e1234c",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de proyectos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Mi Proyecto"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descripción del proyecto"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Crear un nuevo proyecto",
                "operationId": "d91af95183d5934b42aec1325d5165db",
                "requestBody": {
                    "description": "Datos del proyecto a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del proyecto",
                                        "type": "string",
                                        "example": "Mi Nuevo Proyecto"
                                    },
                                    "description": {
                                        "description": "Descripción del proyecto",
                                        "type": "string",
                                        "example": "Descripción detallada del proyecto"
                                    },
                                    "status": {
                                        "description": "Estado del proyecto",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "completed"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Mi Nuevo Proyecto"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descripción detallada del proyecto"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Obtener detalles de un proyecto",
                "operationId": "4c45d4f68ff6ca2929032424e6da2639",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del proyecto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Mi Proyecto"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descripción del proyecto"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "tasks_count": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Actualizar un proyecto",
                "operationId": "546bd69cc1e45c7e4a50c3da131d0969",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del proyecto a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre del proyecto",
                                        "type": "string",
                                        "example": "Proyecto Actualizado"
                                    },
                                    "description": {
                                        "description": "Descripción del proyecto",
                                        "type": "string",
                                        "example": "Nueva descripción del proyecto"
                                    },
                                    "status": {
                                        "description": "Estado del proyecto",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "completed"
                                        ],
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Proyecto Actualizado"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Nueva descripción del proyecto"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Eliminar un proyecto",
                "operationId": "ef93971582acea3d860c54c21c5fd8e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proyecto eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Proyecto eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            }
        },
...
    },
...
}